data_collector/requirements.txt
kafka-python
pyyaml
psycopg2-binary
sqlalchemy